scripts/requirements.txt
PyYAML>=6.0
azure-eventhub>=5.11.0
azure-identity>=1.15.0
websocket-client>=1.6.0
